Change vector division to div_euclid and add remainder via rem_euclid
We use rem_euclid so that the remainder will always be nonnegative.
This helps to implement toroidal manifolds.

To keep the relation between div and rem this requires changing
how division rounds.

Rounding changes from towards zero to rounding down when the divisor
is positive (the typical use case), or rounding up when the divisor
is negative.  In particular, the result of the center method
of a bounding box may change, if the center is negative and
rounding is needed.